FastAPI_Backend/requirements.txt
tensorflow-cpu
fastapi
uvicorn
python-multipart
pillow
matplotlib
numpy
gunicorn==20.1.0
uvloop
httptools
